feat(0250): GREEN — relocate pre-2007 traditions to neutral module
Move the pre-2007 co-citation analysis machinery (build_pre2007_traditions
and its private helpers _load_data, _build_cocitation_network,
_make_node_label, _assign_traditions, plus TOP_N/MIN_COCIT/RANDOM_STATE/
TRADITION_ANCHORS) out of the plot module plot_fig_traditions.py into a new
neutral scripts/_pre2007_traditions.py. Both the plot layer and
compute_null_separation.py now import it from there, severing the last
compute → plot backward arrow that test_layering.py allowlisted.

Pure relocation: the functions and constants are byte-identical, the data
path is unchanged, so the figure and the null-separation output are
identical by construction. plot_fig_traditions.py keeps its rendering
functions and imports RANDOM_STATE (spring_layout seed) and
TRADITION_ANCHORS (_draw_labels) from the new module.

test_layering.py's allowlist is now empty and the guard passes with the
entry removed; test_null_separation.py's arch-rule-9 test repoints to the
neutral module.
